inline templates of mini-cards and lift them into same directory
