--- mineos-app/src/app/components/mini-cards/memory-card/memory-card.component.html
@@ -1,2 +1,0 @@
-<app-mini-card [icon]="faMemory" color="green" [value]="(serverMemory$ | async | bytes_to_mb) ?? 'n/a'" text="RAM Free"></app-mini-card>
-
