--- mineos-app/src/app/components/mini-cards/player-card/player-card.component.html
@@ -1,1 +1,0 @@
-<app-mini-card [icon]="faUsers" [value]="(activePlayers$ | async) ?? 'n/a'" text="Players Online"></app-mini-card>
