+++ mineos-app/src/app/components/mini-cards/server-card.component.ts
@@ -1,13 +1,19 @@
-import {Component, OnDestroy} from '@angular/core';
-import {faCheckCircle} from '@fortawesome/free-solid-svg-icons';
-import {BehaviorSubject, Subscription} from 'rxjs';
-import {ServerHeartbeat} from 'src/app/models/server-heartbeat';
-import {MineosSocketService} from 'src/app/services/mineos-socket.service';
+import { Component, OnDestroy } from '@angular/core';
+import { faCheckCircle } from '@fortawesome/free-solid-svg-icons';
+import { BehaviorSubject, Subscription } from 'rxjs';
+import { ServerHeartbeat } from 'src/app/models/server-heartbeat';
+import { MineosSocketService } from 'src/app/services/mineos-socket.service';
 
 @Component({
   selector: 'app-server-card',
-  templateUrl: './server-card.component.html',
-  styleUrls: [],
+  template: `
+      <app-mini-card [icon]="faCheckCircle"
+                     [value]="(activeServers$ | async) ?? 'n/a'"
+                     color="green"
+                     text="Servers Running">
+      </app-mini-card>
+  `,
+  styleUrls: []
 })
 export class ServerCardComponent implements OnDestroy {
   faCheckCircle = faCheckCircle;
@@ -15,6 +21,7 @@ export class ServerCardComponent implements OnDestroy {
   activeServers$: BehaviorSubject<number> = new BehaviorSubject<number>(0);
   heartbeats: Map<string, Subscription> = new Map<string, Subscription>();
   serverCountMap: Map<string, boolean> = new Map<string, boolean>();
+
   constructor(private mineosSocket: MineosSocketService) {
     this.mineosSocket.serverList().subscribe((serverList) => {
       serverList.forEach((serverName) => {
