--- mineos-app/src/app/components/mini-cards/server-card/server-card.component.html
@@ -1,1 +1,0 @@
-<app-mini-card [icon]="faCheckCircle" color="green" [value]="(activeServers$ | async) ?? 'n/a'" text="Servers Running"></app-mini-card>
